# trust_rnt/src/main/java/com/trust/controller/api/ArticleApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.trust.controller.api;

import com.trust.utils.Result;
import com.trust.utils.ResultData;
import io.swagger.annotations.*;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;

import javax.validation.constraints.NotNull;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-15T03:06:26.991Z")
@Api(value = "article", description = "文章内容相关接口")
public interface ArticleApi {

    @ApiOperation(value = "获取分页内容信息", notes = "分页查询，需要页码，数量及内容类别,通知通告、农业要闻、政策法规、病虫害集分页列表共用.", response = Result.class, tags={ "article", })
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "success", response = Result.class) })
    @RequestMapping(value = "/rnt/article/page",
            produces = { "application/json" },
            method = RequestMethod.GET)
    Result getArticleByPage(@NotNull @ApiParam(value = "页码", required = true) @RequestParam(value = "pageNum", required = true) Integer pageNum, @NotNull @ApiParam(value = "每页显示条数", required = true) @RequestParam(value = "pageSize", required = true) Integer pageSize, @NotNull @ApiParam(value = "分类编码 通知公告：trust_notice  政策法规：trust_policie_regulations 农业要闻：agricultural_news 病虫害集根据病虫害子类编码获取", required = true) @RequestParam(value = "categoryCode", required = true) String categoryCode);




    @ApiOperation(value = "按标题获取分页内容信息", notes = "分页按标题查询，需要页码，数量及内容类别,通知通告、农业要闻、政策法规、病虫害集分页列表共用.", response = Result.class, tags={ "article", })
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "success", response = Result.class) })

    @RequestMapping(value = "/rnt/article/search/page",
            produces = { "application/json" },
            method = RequestMethod.GET)
    Result getArticleByTitlePage(@NotNull @ApiParam(value = "页码", required = true) @RequestParam(value = "pageNum", required = true) Integer pageNum, @NotNull @ApiParam(value = "每页显示条数", required = true) @RequestParam(value = "pageSize", required = true) Integer pageSize, @NotNull @ApiParam(value = "分类编码 通知公告：trust_notice  政策法规：trust_policie_regulations 农业要闻：agricultural_news 病虫害集根据病虫害子类编码获取", required = true) @RequestParam(value = "categoryCode", required = true) String categoryCode,@NotNull @ApiParam(value = "标题", required = true) @RequestParam(value = "title", required = true) String title);



    @ApiOperation(value = "获取内容详情信息", notes = "文章、公告、病虫害、详情查看", response = Result.class, tags={ "article", })
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "success", response = Result.class) })

    @RequestMapping(value = "/rnt/article/details",
            produces = { "application/json" },
            method = RequestMethod.GET)
    Result getArticleInfoById(@NotNull @ApiParam(value = "文章id", required = true) @RequestParam(value = "id", required = true) Long id, @NotNull @ApiParam(value = "访问端类型: 电脑端:PC/移动APP端：Mobile/移动大屏端：Bigscreen", required = true) @RequestParam(value = "sourceType", required = true) String sourceType, Model model);

}