# trust_rnt/src/main/java/com/trust/controller/api/CategoryApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.trust.controller.api;

import com.trust.utils.Result;
import com.trust.utils.ResultData;
import io.swagger.annotations.*;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;

import javax.validation.constraints.NotNull;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-15T03:06:26.991Z")

@Api(value = "category", description = "病虫害分类相关接口")
public interface CategoryApi {


    @ApiOperation(value = "获取分类子类树结构", notes = "病虫害树", response = Result.class, tags={ "category", })
    @ApiResponses(value = {
    @ApiResponse(code = 200, message = "success", response = Result.class) })
    
    @RequestMapping(value = "/rnt/category/list",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    Result getCategoryTreeList(@NotNull @ApiParam(value = "分类编码 病虫害：trust_insect_pests", required = true) @RequestParam(value = "categoryCode", required = true) String categoryCode);

}