# trust_rnt/src/main/java/com/trust/controller/api/PricequotesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.trust.controller.api;

import com.trust.utils.Result;
import com.trust.utils.ResultData;
import io.swagger.annotations.*;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-15T03:06:26.991Z")

@Api(value = "pricequotes", description = "价格行情相关接口")
public interface PricequotesApi {

    @ApiOperation(value = "获取价格行情", notes = "此版本暂无需传入参数后期多系统可能需传入系统编码城市编码 categoryCode是动态的分类编码  latest标识不同种类最新的", response = Result.class, tags={ "priceQuotes", })
    @ApiResponses(value = {
        @ApiResponse(code = 200, message = "success", response = Result.class) })
    
    @RequestMapping(value = "/rnt/pricequotes/list",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    Result getPriceQuotesByparams(@ApiParam(value = "城市编码") @RequestParam(value = "city", required = false) String city, @ApiParam(value = "系统编码") @RequestParam(value = "sysCode", required = false) String sysCode);

}