# trust_rnt/src/main/java/com/trust/controller/api/RntversionApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.trust.controller.api;

import com.trust.utils.Result;
import com.trust.utils.ResultData;
import io.swagger.annotations.*;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;

import javax.validation.constraints.NotNull;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-15T03:06:26.991Z")

@Api(value = "rntversion", description = "版本信息相关接口")
public interface RntversionApi {

    @ApiOperation(value = "根据系统名称获取版本信息", notes = "系统名称.", response = Result.class, tags={ "rntVersion", })
    @ApiResponses(value = {
        @ApiResponse(code = 200, message = "success", response = Result.class) })
    
    @RequestMapping(value = "/rnt/rntversion/detail",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResultData getRntVersionInfo(@NotNull @ApiParam(value = "系统名称 IOS\\Android\\Backstage", required = true) @RequestParam(value = "sysName", required = true) String sysName, @ApiParam(value = "系统编码：本版本暂无此参数后期多系统或需传入此参数") @RequestParam(value = "sysCode", required = false) String sysCode);

}